requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
groq>=0.12.0
PyPDF2>=3.0.1
python-docx>=0.8.11
python-multipart>=0.0.6
numpy>=1.24.3
pydantic>=2.6.0
pydantic-settings>=2.2.0
requests>=2.31.0
python-json-logger>=2.0.4
python-dotenv>=1.0.0
httpx>=0.26.0
